1.21.5

+++ fabric/src/main/java/com/jahirtrap/walljump/logic/DoubleJumpLogic.java
@@ -39,7 +39,7 @@ public class DoubleJumpLogic {
                 jumpCount--;
 
                 pl.resetFallDistance();
-                ClientPlayNetworking.send(new MessageFallDistance(pl.fallDistance));
+                ClientPlayNetworking.send(new MessageFallDistance((float) pl.fallDistance));
             }
 
             jumpKey = true;
